requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.9
cloudinary==1.28.0
dj-database-url==0.5.0
Django==3.2.9
django-bootstrap4==21.1
django-cloudinary-storage==0.3.0
django-registration-redux==2.9
djangorestframework==3.12.4
idna==3.3
Pillow==8.4.0
psycopg2-binary==2.9.2
python-decouple==3.5
pytz==2021.3
requests==2.26.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
urllib3==1.26.7
